[Enhancement #153] Support pagination for last edited assets on dashboard.
