+++ src/component/dashboard/widget/lastcommented/MyLastCommentedAssets.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import CommonLastCommentedAssets from "./CommonLastCommentedAssets";
-import { loadMyLastCommented } from "../../../../action/AsyncCommentedAssetActions";
+import { loadMyLastCommented } from "../../../../action/AsyncDashboardActions";
 
 const MyLastCommentedAssets: React.FC = () => (
   <CommonLastCommentedAssets loadAssets={loadMyLastCommented} />
