+++ src/component/dashboard/widget/lastedited/MyLastEditedAssets.tsx
@@ -1,18 +1,9 @@
-import * as React from "react";
-import { connect } from "react-redux";
-import { ThunkDispatch } from "../../../../util/Types";
-import { loadMyAssets } from "../../../../action/AsyncActions";
-import RecentlyModifiedAsset from "../../../../model/RecentlyModifiedAsset";
+import React from "react";
+import { loadMyAssets } from "../../../../action/AsyncDashboardActions";
 import CommonLastEditedAssets from "./CommonLastEditedAssets";
 
-interface AllLastEditedAssetsProps {
-  loadAssets: () => Promise<RecentlyModifiedAsset[]>;
-}
+export const MyLastEditedAssets: React.FC = () => (
+  <CommonLastEditedAssets loadAssets={loadMyAssets} />
+);
 
-export const MyLastEditedAssets: React.FC<AllLastEditedAssetsProps> = (
-  props
-) => <CommonLastEditedAssets {...props} />;
-
-export default connect(undefined, (dispatch: ThunkDispatch) => ({
-  loadAssets: () => dispatch(loadMyAssets()),
-}))(MyLastEditedAssets);
+export default MyLastEditedAssets;
